move MessageTypeMapInitializer to Beckett project, drop Beckett.Testing project / package
