+++ .github/workflows/publish.yml
@@ -33,8 +33,6 @@ jobs:
         run: dotnet pack src/Beckett/Beckett.csproj --no-build --configuration Release --output ${{ env.NuGetDirectory }}
       - name: Create Beckett.Dashboard NuGet Package
         run: dotnet pack src/Beckett.Dashboard/Beckett.Dashboard.csproj --no-build --configuration Release --output ${{ env.NuGetDirectory }}
-      - name: Create Beckett.Testing NuGet Package
-        run: dotnet pack src/Beckett.Testing/Beckett.Testing.csproj --no-build --configuration Release --output ${{ env.NuGetDirectory }}
       - name: Publish NuGet Package
         run: |
           foreach($file in (Get-ChildItem "${{ env.NuGetDirectory }}" -Recurse -Include *.nupkg)) {
