+++ Beckett.sln
@@ -16,8 +16,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Beckett.Dashboard", "src\Be
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "TaskHub", "TaskHub", "{EF59B95F-62EA-40AC-9F27-121138C1D031}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Beckett.Testing", "src\Beckett.Testing\Beckett.Testing.csproj", "{238A0F1F-0B14-4F53-92C1-A32488A8A96B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Beckett.SourceGenerators", "src\Beckett.SourceGenerators\Beckett.SourceGenerators.csproj", "{F75D8D08-1665-4C89-9BF3-775DB452C084}"
 EndProject
 Global
@@ -50,10 +48,6 @@ Global
 		{16B9AAA4-7DD5-4DA8-BAC6-E6885E18410A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{16B9AAA4-7DD5-4DA8-BAC6-E6885E18410A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{16B9AAA4-7DD5-4DA8-BAC6-E6885E18410A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{238A0F1F-0B14-4F53-92C1-A32488A8A96B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{238A0F1F-0B14-4F53-92C1-A32488A8A96B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{238A0F1F-0B14-4F53-92C1-A32488A8A96B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{238A0F1F-0B14-4F53-92C1-A32488A8A96B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F75D8D08-1665-4C89-9BF3-775DB452C084}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F75D8D08-1665-4C89-9BF3-775DB452C084}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F75D8D08-1665-4C89-9BF3-775DB452C084}.Release|Any CPU.ActiveCfg = Release|Any CPU
