--- src/Beckett.Testing/StateSpecificationFixture.cs
@@ -1,6 +1,0 @@
-namespace Beckett.Testing;
-
-public abstract class StateSpecificationFixture<T> where T : IApply, new()
-{
-    protected StateSpecification<T> Specification => new();
-}
