+++ src/Beckett/Messages/MessageTypeMapInitializer.cs
@@ -2,10 +2,16 @@ using System.Reflection;
 using Beckett.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Beckett.Testing;
+namespace Beckett.Messages;
 
 public static class MessageTypeMapInitializer
 {
+    /// <summary>
+    /// Initialize the <see cref="MessageTypeMap"/> using <see cref="IBeckettModule"/> implementations
+    /// discovered in the provided list of assemblies. Mostly useful for global test setup for unit tests, however it
+    /// can be used as an alternative means of setting up the system where appropriate.
+    /// </summary>
+    /// <param name="assemblies"></param>
     public static void Initialize(params Assembly[] assemblies)
     {
         var builder = new BeckettBuilder(new ServiceCollection());
